fix: cast SANCTUM_TOKEN_EXPIRY_MINUTES to int; restore config in test
- cast env() result to (int) so config('sanctum.expiration') always
  returns int regardless of whether the env var is a string from .env
- rename test to 'token expiration defaults to 1440 minutes' to match
  what it actually asserts instead of implying broader configurability
- capture and restore config('sanctum.expiration') after mutation in
  'expired bearer tokens are rejected' so test order is irrelevant

+++ config/sanctum.php
@@ -43,7 +43,7 @@ return [
     |
     */
 
-    'expiration' => env('SANCTUM_TOKEN_EXPIRY_MINUTES', 1440),
+    'expiration' => (int) env('SANCTUM_TOKEN_EXPIRY_MINUTES', 1440),
 
     /*
     |--------------------------------------------------------------------------
